Added cd command.

+++ bootloader/stage2/disk.c
@@ -17,14 +17,15 @@ bool diskInitialize(DISK* disk, uint8_t driveNumber) {
 }
 
 void lbaToChs(DISK* disk, uint32_t lba, uint16_t* sectors, uint16_t* cylinders, uint16_t* heads) {
-    *sectors = lba % disk->sectors + 1;
-    *heads = (lba / disk->sectors) % disk->heads;
-    *cylinders = (lba / disk->sectors) / disk->heads;
+    *sectors = (uint16_t)lba % disk->sectors + 1;
+    *heads = (uint16_t)lba / disk->sectors % disk->heads;
+    *cylinders = (uint16_t)lba / disk->sectors / disk->heads;
 }
 
 bool diskReadSectors(DISK* disk, uint32_t lba, uint8_t count, void far* outData) {
     uint16_t sectors, cylinders, heads;
     int i;
+
     lbaToChs(disk, lba, &sectors, &cylinders, &heads);
 
     for (i = 0; i < 3; i++) {
